Revert "Revert "version 0.3.20""
This reverts commit b9a9d25c75be96abc2a0a541d1b562f5428b5296.

+++ application/Core/Modules/Cache/CacheCore.php
@@ -43,7 +43,7 @@ class CacheCore
 		return $this->dirName;
 	}	
 	
-	public function store($fname, $data, $exp = null)
+	public function store($fname, $data, $exp = 0)
 	{
 		if(!$this->cache) return null;
 		if(!isset($exp))
@@ -111,11 +111,18 @@ class CacheCore
 		{
 			foreach($list as $k => $v)
 			{
-				if($v < time())
+				if($v != 0 && $v < time())
 				{	
-					if(!unlink($k . $this->extension))
-						throw new \Exception('cannot delete cache file'.$k);
-					unset($list[$k]);
+					if(file_exists($k))
+					{
+						if(!unlink($k . $this->extension))
+							throw new \Exception('cannot delete cache file'.$k);
+						unset($list[$k]);
+					}
+					else
+					{
+						unset($list[$k]);
+					}
 				}
 			}
